api/requirements.txt
fastapi>=0.115.0
uvicorn>=0.30.0
python-multipart>=0.0.12
onnxruntime>=1.18.0
numpy>=1.26.0
Pillow>=10.0.0
